Refactored code by passing some code logic to contact component leading to reducing some lines of code and a nice structure
